src/task_worker/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
debugpy==1.6.3
Deprecated==1.2.13
future==0.18.2
kombu==5.2.4
packaging==21.3
Pillow==9.3.0
prompt-toolkit==3.0.32
pyparsing==3.0.9
pytz==2022.6
PyYAML==6.0
redis==4.3.4
six==1.16.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
